feat(lib): rename `models` into `builtins`
